# apps/go_backend/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  categories.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  categories.CategoryPageResponse:
    properties:
      has_more:
        type: boolean
      items:
        items:
          $ref: '#/definitions/categories.Category'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  categories.CreateRequest:
    description: Request payload for creating a category
    properties:
      color:
        example: '#3B82F6'
        maxLength: 50
        minLength: 1
        type: string
      name:
        example: Work
        maxLength: 100
        minLength: 1
        type: string
    required:
    - color
    - name
    type: object
...
  errors.ErrorCode:
    enum:
    - BAD_REQUEST
    - VALIDATION_FAILED
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - CONFLICT
    - TOO_MANY_REQUESTS
    - INTERNAL_ERROR
    - DATABASE_ERROR
    - SERVICE_UNAVAILABLE
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeValidationFailed
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeTooManyRequests
    - CodeInternalError
    - CodeDatabaseError
    - CodeServiceError
...
  response.APIError:
    properties:
      code:
        $ref: '#/definitions/errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
    type: object
  response.OperationMessage:
    properties:
      message:
        type: string
    type: object
  tasks.CreateRequest:
    description: Request payload for creating a task
    properties:
      category_id:
        example: categories:work123
        type: string
      end_date:
        example: "2025-11-25T17:00:00Z"
        type: string
      journal:
        example: Capture high-level goals
        maxLength: 10000
        type: string
      negatives:
        example:
        - Some distractions
        items:
          type: string
        type: array
      note:
        example: Focus on top priorities
        maxLength: 5000
        type: string
      positives:
        example:
        - Great progress
        - In flow
        items:
          type: string
        type: array
      priority:
        example: 1
        maximum: 100
        minimum: -100
        type: integer
      source:
        example: manual
        type: string
      start_date:
        example: "2025-11-24T09:00:00Z"
        type: string
      title:
        example: Plan tomorrow
        maxLength: 500
        minLength: 1
        type: string
    required:
    - end_date
    - start_date
    - title
    type: object
  tasks.Task:
    properties:
      category:
        $ref: '#/definitions/categories.Category'
      completed:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      journal:
        type: string
      negatives:
        items:
          type: string
        type: array
      note:
        type: string
      positives:
        items:
          type: string
        type: array
      priority:
        type: integer
      source:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  tasks.TaskPageResponse:
    properties:
      has_more:
        type: boolean
      items:
        items:
          $ref: '#/definitions/tasks.Task'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  tasks.UpdateRequest:
    description: Request payload for updating a task
    properties:
      category_id:
        type: string
      completed:
        type: boolean
      end_date:
        type: string
      journal:
        maxLength: 10000
        type: string
      negatives:
        items:
          type: string
        type: array
      note:
        maxLength: 5000
        type: string
      positives:
        items:
          type: string
        type: array
      priority:
        maximum: 100
        minimum: -100
        type: integer
      start_date:
        type: string
      title:
        maxLength: 500
        minLength: 1
        type: string
    type: object
info:
  contact: {}
  description: Daily journal backend powered by SurrealDB.
  title: Lucid Logs API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      description: Get paginated categories for the authenticated user
      parameters:
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.CategoryPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a category for organizing tasks
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categories.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/categories.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - categories
...
  /api/v1/tasks:
    get:
      description: Get paginated list of tasks for the authenticated user
      parameters:
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.TaskPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task
      parameters:
      - description: Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tasks.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      description: Soft delete a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OperationMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a single task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.Task'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tasks.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"